refactor: move analytics code from ui library
