+++ apps/cow-fi/lib/hooks/useAddRpcWithTimeout.ts
@@ -1,5 +1,4 @@
 import { useCallback } from 'react'
-// import { addRpcUrlAction as sendToAnalytics, AddRpcUrlActionType } from '@src/lib/analytics/events'
 import { ERROR_ADD_MANUALLY_MESSAGE } from '@/util/handleRpcError'
 import { useAddRpcEndpoint } from './useAddRpcEndpoint'
 import { WalletClient } from 'viem'
