+++ libs/analytics/src/index.ts
@@ -3,5 +3,7 @@ export { WebVitalsAnalytics } from './webVitals/WebVitalsAnalytics'
 export { initPixelAnalytics } from './pixels/initPixelAnalytics'
 export { CowAnalyticsGoogle } from './googleAnalytics/CowAnalyticsGoogle'
 export { initCowAnalyticsGoogle } from './googleAnalytics/initGoogleAnalytics'
+export { useAnalyticsReporter } from './hooks/useAnalyticsReporter'
 export * from './context/CowAnalyticsContext'
 export * from './CowAnalytics'
+export * from './events'
