+++ libs/ui/src/index.ts
@@ -4,9 +4,6 @@ export * from './enum'
 export * from './types'
 export * from './consts'
 export * from './theme'
-
-export { useAnalyticsReporter } from './analytics/useAnalyticsReporter'
-
 export * from './pure/ButtonSecondaryAlt'
 export * from './pure/Button'
 export * from './pure/CenteredDots'
